// client/src/generated/Ev.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Ev = { "t": "Dmg", x: number, y: number, v: number, crit: boolean, p: boolean, } | { "t": "Heal", x: number, y: number, v: number, } | { "t": "Boom", x: number, y: number, r: number, k: number, } | { "t": "Died", id: number, x: number, y: number, kind: number, } | { "t": "Raise", x: number, y: number, } | { "t": "Immune", x: number, y: number, } | { "t": "Tile", x: number, y: number, v: number, } | { "t": "Msg", text: string, } | { "t": "Coins", x: number, y: number, v: number, };